+++ src/__tests__/notifications/notifications-parser.test.ts
@@ -14,6 +14,10 @@ import { ApiKeyExpiredMock, ApiKeyExpiredMockExpectation } from '../mocks/notifi
 import { ApiKeyRevokedMock, ApiKeyRevokedMockExpectation } from '../mocks/notifications/api-key-revoked.mock.js';
 import { ApiKeyExpiringMock, ApiKeyExpiringMockExpectation } from '../mocks/notifications/api-key-expiring.mock.js';
 import { ApiKeyUpdatedMock, ApiKeyUpdatedMockExpectation } from '../mocks/notifications/api-key-updated.mock.js';
+import {
+  ApiKeyExposureCreatedMock,
+  ApiKeyExposureCreatedMockExpectation,
+} from '../mocks/notifications/api-key-exposure-created.mock.js';
 import { BusinessCreatedMock, BusinessCreatedMockExpectation } from '../mocks/notifications/business-created.mock.js';
 import { BusinessUpdatedMock, BusinessUpdatedMockExpectation } from '../mocks/notifications/business-updated.mock.js';
 import {
@@ -159,6 +163,7 @@ describe('Notifications Parser', () => {
     [ApiKeyExpiringMock.event_type, ApiKeyExpiringMock, ApiKeyExpiringMockExpectation],
     [ApiKeyRevokedMock.event_type, ApiKeyRevokedMock, ApiKeyRevokedMockExpectation],
     [ApiKeyUpdatedMock.event_type, ApiKeyUpdatedMock, ApiKeyUpdatedMockExpectation],
+    [ApiKeyExposureCreatedMock.event_type, ApiKeyExposureCreatedMock, ApiKeyExposureCreatedMockExpectation],
     [BusinessCreatedMock.event_type, BusinessCreatedMock, BusinessCreatedMockExpectation],
     [BusinessUpdatedMock.event_type, BusinessUpdatedMock, BusinessUpdatedMockExpectation],
     [BusinessImportedMock.event_type, BusinessImportedMock, BusinessImportedMockExpectation],